\name{R-makemymap-package}
\alias{R-makemymap-package}
\alias{R-makemymap}
\docType{package}
\title{
    Useful functions for handling maps with ggplot2
}
\description{
    This package contains some wrapper function enabling the user to
    access GSHHS shorelines and plot these at various map projections
    using \link{ggplot}, \link{proj} and \link{maptools}.
}
\details{
    \tabular{ll}{
Package: \tab R-makemymap\cr
             Version: \tab 0.2\cr
             Date: \tab 2014-02-14\cr
             Depends: \tab \cr
             Imports: \tab maptools proj rgeos\cr
             Suggests: \tab \cr
             License: \tab GPL version 2 or newer\cr
             URL: \tab http://www.r-project.org, http://saf.met.no\cr
             Built: \tab R 2.3.1; i686-redhat-linux-gnu; 2006-10-02 12:10:01; unix\cr
    }

Index:
    \preformatted{
    }
}
\author{
    Øystein Godøy <o.godoy@met.no>

    Maintainer: Øystein Godøy <o.godoy@met.no>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
